web-version edit & HTML minor edit
I know I shouldn't have, but I moved to `/var/www/public/`. This stopped the version from reporting for admin LTE.

I'm actually planning on moving it again and this has helped as `.git` repository is in the same folder as `footer.php` so you only need to worry about specifying path if that changes.