Make vertical axis start from 0
Main goal was to change the Y-axis to start from 0 as starting anywhere else could distort the data too much. Good practice is to always start from 0, unless you have a really good reason not to.
Besides that I removed "var monthNames" which was not used anywhere and "jsonData" was assigned but the referenced variable was never used.
In addition, I just beautified the code to make it more readable and structured.