Add Quad9 DNS Servers to DNS Servers List
I've added the Quad9 DNS servers (IPv4 and IPv6) to the DNS Servers list.  (https://www.quad9.net/)  I've opted not to also add their non-secure alternate servers, as those mainly exist for diagnostics of false positives, and are not the primary focus of the Quad9 service.  They do not offer secondary servers, as their primary servers are already geographically distributed and redundant.