Update dashboard for better fit of small boxes
On the raspberrypi official touchscreen (and other screens that fit "small" and "medium"), the small boxes take up the 12 columns which forces to scroll and is generally ineffective as a quick glance dashboard. I have updated to col-sm-6 so there are 2 boxes on 2 rows so now you can see all those stats plus the queries last 24 without having to scroll. the col-xs-12 is not necessary as it will default to that.

Signed-off-by: Derek Bowes <db@bowesmail.com>