Prevent PHP Fatal error:  Maximum execution time of 30 seconds exceeded
I have a different service running on FTL default port and so FTL service was not properly started. I searched quite a long time to discover why the Dashboard never loaded and finally found the reason.

This patch counts empty lines and dies properly with a JSON error. This fixed my issue.