Fix TypeError if no extended DNS error is available
If no extended DNS errors are available, the "ede" variable is assigned an undefined array value, hence "ede" is undefined. Since the length method is not available on undefined variables, a TypeError is thrown later in the script, which breaks the DNS queries view in the admin panel.

This commit solves the issue by assigning an empty string, if no EDE is available (data[11] is undefined).

Signed-off-by: MichaIng <micha@dietpi.com>