Background colors on Query log + striped table
Improves PR #1893

The CSS for the background was created using "!important", to force the background colors for blocked/allowed on the table lines, but this introduced a visual problem: all lines have the same backgrounds. No more striped tables.

To solve this, I moved the effect from the `tr` tags to `td` tags. Now the rows (tr) have striped backgrounds and the cells adds the color layer.

Signed-off-by: rdwebdesign <github@rdwebdesign.com.br>