Update api.php?setTempUnit function to require authentication
There was a lack of a check to the $auth variable. This allowed any unauthenticated user to change the temperature unit of the pi-hole application. Adding the check of this variable fixes this issue.

The absence of authentication of this function was deemed not a security risk but maintainers were open to a potential merge depending on the release of v6.

Report: https://github.com/pi-hole/web/security/advisories/GHSA-mffw-5jmg-9wjq

Signed-off-by: kiyell <kysoftware@gmail.com>